refactor: extract shared Lean source parser
